add: fases, pontuação, inimigo valoz, inimigo perseguidor, chefão e tiros

+++ TopGear/Teste/src/com/tutorial/main/GameObject.java
@@ -5,11 +5,11 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {  //objetos de jogo
     
-    protected int x, y; //x,y no plano cartesiano
+    protected float x, y; //x,y no plano cartesiano
     protected ID id;    //id para diferenciar jogador e inimigos
-    protected int velX, velY;
+    protected float velX, velY;
 
-    public GameObject(int x, int y, ID id){
+    public GameObject(float x, float y, ID id){
         this.x = x;
         this.y = y;
         this.id = id;
@@ -20,17 +20,17 @@ public abstract class GameObject {  //objetos de jogo
     public abstract Rectangle getBounds();      //adiciona bordas aos retangulos para colisão
 
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
-    public int getX() {
+    public float getX() {
         return this.x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }  
-    public int getY() {
+    public float getY() {
         return this.y;
     }
 
@@ -41,17 +41,17 @@ public abstract class GameObject {  //objetos de jogo
         return this.id;
     }
 
-    public void setVelX(int velX) {
+    public void setVelX(float velX) {
         this.velX = velX;
     }
-    public int getVelX() {
+    public float getVelX() {
         return this.velX;
     }
     
-    public void setVelY(int velY) {
+    public void setVelY(float velY) {
         this.velY = velY;
     }    
-    public int getVelY() {
+    public float getVelY() {
         return this.velY;
     }
